Eu te amo Carlos Gustavo Cecyn Lundgren!!
Obrigado Mummi!! Feliz 62 anos!!

at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU 4 core 8 threads @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 8 Pro - with Portuguese - Brazil- 10Mpbs Download and 2Mpbs Upload NET Virtua Internet with LG 23´´ monitor, Microsoft Comfort Curve 2000 v1.0 Wired Keyboard and Microsoft Comfort Optical Wired Mouse 3000 at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

Obrigado Carol, Lucinho e Lúcio!!

Obrigado Célia!!

Obrigado Papa Francisco!!

Obrigado Ralf Sieves!!

Obrigado Rafaal Cecyn Lundren!!

-∞
Tack Sverige!!
{
  Tack Carlos Gustavo Cecyn Lundgren!!
  Tack Léia Maheli de Souza Tsumanuma!! Mummi
  Tack Simon Petrus Lundgren!!
  Tack Iara Cecyn Lundgren!!
  Tack Gabrielli Cecyn Lundgren!! Gabi
  Tack Rafael Cecyn Lundgren!!
  Tack Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
日本ありがとうございます
｛
　　　日歩路主ありがとうございます！！


   -----
　　　妻沼国男ありがとうございます！！
   -----


　　　日歩路主ありがとうございます！！
｝

Gracias España
{
   Gracias Carlos Carrasco Barriga!!
}

Dank Österreich
{
   Dank Vincent Rijmen!!
   Dank www.iaik.tugraz.at!!
   Dank Schinagl at http://schinagl.priv.at/!! - Link Shell Extension
}

Obrigado Cits
{
   Obrigado Rejeane!!
   Obrigado Monique!!
   Obrigado Gilmar Tuchinski!!
   Obrigado Marcos Machado!!
   Obrigado Camila Napoli Madureira Bührer!!
}

Obrigado Italia-Argentina
{
	Obrigado Papa Francisco!!
}

Obrigado Brasil
{
   Obrigado Católicos do Brasil!!
   Obrigado Santos!!
   Obrigado Pessoal Da Reforma da Pizzeria aqui ao Lado Solloviaggio!!
   Obrigado Pessoal Da Pizzeria Solloviaggio!!
   Obrigado vizinhos da Alameda Princesa Izabel, 3000 :!!
     Obrigado Gil!!
     Obrigado dono mulato de Gol escuro!!
     Obrigado filho do dono mulato de Gol escuro!!
     Obrigado dono branco de Gol prata e família!!
   Obrigado funcionários da CEF do CEFET!!
   Obrigado funcionários do CEFET!! Obrigado CEFET!! Obrigado pessoal do CEFET!!
   Obrigado dois meninos que reformaram Armário Carol Santos Canal 2 2013-03-23 caprichado e bonitinhos!!
   Obrigado meninos trabalhadores do Brasil!!
   -----
   Obrigado Cintia do Winnikes!!                    _-\
   Obrigado Alisson do Winnikes!!                      |
   Obrigado Ótica Winnikes do Shopping Curitiba!!      |
   Obrigado Shopping Curitiba!!                      -/
   Thank you posto Isejima o Fazendeiro!!
   -----
   Obrigado guardas!!
   Obrigado Thomas Friesen!!
   -----
   Obrigado moças e mulheres...
   Obrigado moços e homens...
   Obrigado Keila Mª da Silva!! Babysitter do Lucinho
   Obrigado Rosélia da Kadu Imóveis!!
   Obrigado Márcia da Kadu Imóveis!!
   Obrigado Sílvia da Kadu Imóveis!!
   Obrigado Ivete - irmã da Silvia da Kadu!!
   -----
   Obrigado Dr. Tassi!!
   Obrigado Margarida M. Gomes de Oliveira!!
   Obrigado Shirley Sirlene Adams!!
   Obrigado Solange!!
   Obrigado CPM - Centro Psiquiátrico Metropolitano!!
   -----
   Obrigado Paulo S. L. M. Barreto (Whirlpool)!!
   Obrigado USP!!
   -----
   Obrigado Meu Deus God Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Deus!!
   -----
   Obrigado Nizia Lupinacci!!
   Obrigado Kleberson Chita Massaro Rodrigues!!
   Obrigado Alessandro Vargas Pellanda!!
   Obrigado Fábio Dauer da Silva!!
   Obrigado Fabrício Cavassin Mereniuk!!
   Obrigado Leonardo Fischer!!
   Obrigado Telmo Friesen!!
   Obrigado Sector 8J!!
   Obrigado Carlos Gustavo Cecyn Lundgren Tsumanuma!!
   ------
   Obrigado Rafael Cecyn Lundgren!!
   ------
   Obrigado Meu Senhor Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Espírito Santo Carlos Gustavo Cecyn Lundgren!!
   Obrigado Minha Deusa Léia Maheli de Souza Tsumanuma""
   ------
   Obrigado Lúcio Flávio Marques Ribeiro!!
   Obrigado Luís Kunio Tsumanuma!!
   Obrigado Lúcio Flávio Tsumanuma Ribeiro!!
   Obrigado Carolina Kiyoko Tsumanuma Ribeiro!!
}


Obrigado CEFET-PR S11-97/2!!
{
   Obrigado Fábio Rizental Coutinho!!
   Obrigado Verônica Isabela Quandt & Edras Pacola!! 
}

Dank Deutschland!!
{
   Dank Jochen Berger - from serverloft!!
   Dank Thomas Strohe - from serverloft!!
   Dank Wolfgang Strunk!!
   Dank Deustchland Jörg!!
   Dank Deustchland Manfred Riedel!!
   Dank Uwe Hermanns!!
   Dank Telmo Friesen!!
   Dank Cycos!!
   Dank Siemens Enterprise!!
   Dank Siemens!!
   Dank SEN R&D A!!
   Dank Sector8J!!
}

Thank you Cycos!!
{
   Thank you Uwe Hermanns!!
   Thank you Telmo Friesen!!
   Thank you Communications!!
   Thank you Cycos!!
}

Thank you United Kingdom!!
{
   Thank you David Beckham!!
}

Thank you United States of America!!
{
   Thank you se-pl-sy-Carlos Gustavo Cecyn Lundgren!!
}

Dziękuję Polska!!
{
   Carlos Gustavo Cecyn Lundgren dziękuję!!
}

Thank you sector 8J!!
{
   Thank you Anderson Silva!!
   Thank you Pietro!!
   Thank you Thiago Ferreira!! - de 1996 do grupo do Cleiton, Anderson Silva e Pietro
   Thank you Thiago Hilger!!
   Thank you César Augusto Serenato!!
   Thank you Bárbara Martinazzo!!
   Thank you Alessando Ledra Zagheni!!
   Thank you Luiz Eduardo de Christo!!
   Thank you Maurício Cunico!!
   Thank you Galileu Terada!!
   Thank you Rodrigo Santos aka Kenshin!!
   Thank you Júlio Henrique Morimoto!!
   Thank you Natascha Ligocki!!
   Thank you Ricardo Henrique Folador!!
   Thank you KS, Jefferson Koji Sato!!
   Thank you Josué Andrade Gomes!!
   Thank you Jorge Alexandre Hastreiter!!
   Thank you Joice Braga Keskoski Modena!!
   Thank you Bruno Leite Reis!!
   Thank you Fernando Belo!!
   Thank you Felipe Calliari Ribas!!
   Thank you Marcos Paulo Riki Ianase!!
   Thank you Jefferson Luiz Dalavechia!!
   Thank you Ênio Toshihiro Ianase!!
   Thank you Alexsandro Vargas Pellanda!!
   Thank you Cleiton Lenkiu!!
   Thank you Fabrício Cavassin Mereniuk!!
   Thank you Jaquinei de Oliveira!!
   Thank you Cleber Jaizer Gomes Campos!!
   Thank you Fábio Dauer da Silva!!
   Thank you Roberto de Melo Gonçalves!! Chatuba!!
   Thank you Marcus Vinícius Silvério!!
   Thank you Rafael Taizo Yamashita!! Japs!!
   Thank you Rafael Alessi Muntsch!!
   Thank you Rafael Neumann de Menezes!!
   Thank you Ricardo Carvalho dos Santos!! Tom Zé!! Tonny Zé!! Toze!!
   Thank you Leonardo Fischer!! Irmaun!!
   Thank you João Miguel Martynyszyn!!
   Thank you Leonardo da Silva Camargo!!
   Thank you Telmo Friesen!!
   Thank you sector 8J!!
   Thank you Carlos Gustavo Cecyn Lundgren!!
}

Thank you sector Joice/Roman!!
{
   Thank you Rafael Neumman de Menezes!!
   Thank you Thiago Tibilletti!!
   Thank you Thiago Pereira!!
}

+∞

-∞
Obrigado Família Souza
{
   Obrigado Isabel de Souza e Santos!!
   Obrigado Dilma Souza!!
}
+∞

-∞
T Obrigado Amigo e Amiga, você é legal!!
1 Obrigado Amigo e Amiga, você é bom!!
2 Obrigado Amigo e Amiga, eu gosto de você!!
n Obrigado Amigo e Amiga, eu gosto muito de você!!
n Obrigado Amigo e Amiga, você é demais!!
+∞

-∞
Obrigado Minha Deusa!!
Obrigado Minha Deusa e Meu Deus - Mummi e Carlos!!
Obrigado Mummi, eu te amo demais!!
Thank you My Goddess Léia Maheli de Souza Tsumanuma!!
+∞

-∞
01 Obrigado Carlinhos, eu te amo demais!!
02 Thank you My Beloved Lover Carlos Gustavo Cecyn Lundgren!!
03 Thank you My Beloved Friend Carlos Gustavo Cecyn Lundgren!!
04 Thank you My Beloved Husband Carlos Gustavo Cecyn Lundgren!!
05 Thank you My Spouse Carlos Gustavo Cecyn Lundgren!!
06 Thank you My Jesus Carlos Gustavo Cecyn Lundgren!!
07 Thank you My Saint Mary My Mother Carlos Gustavo Cecyn Lundgren!!
08 Thank you My Goddess Carlos Gustavo Cecyn Lundgren!!
09 Thank you My God Carlos Gustavo Cecyn Lundgren!!
10 Thank you My Lord Carlos Gustavo Cecyn Lundgren!!
11 Thank you My Holy Spirit Carlos Gustavo Cecyn Lundgren!!
+∞

∞-∞-∞∞-∞-∞
----------
8nnnnnnnn8
++++++++++
∞+∞+∞∞+∞+∞

-∞
Thank you My God!!
+∞

-∞
(En-us) Thank you, (pt-br) eternamente *obrigado* - por veri absoluta infinita complexíssima oposição a desobrigado.
+∞

+++ appseed/core/net/sockets/base/sockets_base_socket_handler.h
@@ -69,13 +69,13 @@ namespace sockets
             Handler().LogError(this, "OnRead", 0, "data on hibernating socket", ::core::log::level_fatal);
             SetCloseAndDelete();
          }
-         void OnOptions(int32_t,int32_t,int32_t,SOCKET) {}
+         void OnOptions(int,int,int,SOCKET) {}
 
       };
 
    public:
 
-      int32_t m_iSelectErrno;
+      int m_iSelectErrno;
 
 
    public:
@@ -89,7 +89,7 @@ namespace sockets
       virtual void RegStdLog(StdLog *log) = 0;
 
       /** Log error to log class for print out / storage. */
-      virtual void LogError(base_socket *p,const string & user_text,int32_t err,const string & sys_err, ::core::log::e_level elevel = ::ca_get_level_warning()) = 0;
+      virtual void LogError(base_socket *p,const string & user_text,int err,const string & sys_err, ::core::log::e_level elevel = ::ca_get_level_warning()) = 0;
 
       // -------------------------------------------------------------------------
       // socket stuff
@@ -106,11 +106,11 @@ namespace sockets
       /** Set read/write/exception file descriptor sets (fd_set). */
       virtual void Set(SOCKET s,bool bRead,bool bWrite,bool bException = true) = 0;
       /** Wait for events, generate callbacks. */
-      virtual int32_t Select(int32_t sec, int32_t usec) = 0;
+      virtual int Select(long sec,long usec) = 0;
       /** This method will not return until an event has been detected. */
-      virtual int32_t Select() = 0;
+      virtual int Select() = 0;
       /** Wait for events, generate callbacks. */
-      virtual int32_t Select(struct timeval *tsel) = 0;
+      virtual int Select(struct timeval *tsel) = 0;
 
       /** Check that a socket really is handled by this socket handler. */
       virtual bool Valid(socket *) = 0;
@@ -127,7 +127,7 @@ namespace sockets
       // Connection pool
       // -------------------------------------------------------------------------
       /** find available open connection (used by connection pool). */
-      virtual PoolSocket * FindConnection(int32_t type,const string & protocol,::net::address_sp&) = 0;
+      virtual PoolSocket * FindConnection(int type,const string & protocol,::net::address_sp&) = 0;
 
       /** Enable connection pool (by default disabled). */
       virtual void EnablePool(bool = true) = 0;
@@ -172,11 +172,11 @@ namespace sockets
       /** Queue a dns request.
       \param host Hostname to be resolved
       \param port Port number will be echoed in socket::OnResolved callback */
-      virtual int32_t Resolve(socket *,const string & host,port_t port) = 0;
-      virtual int32_t Resolve6(socket *,const string & host,port_t port) = 0;
+      virtual int Resolve(socket *,const string & host,port_t port) = 0;
+      virtual int Resolve6(socket *,const string & host,port_t port) = 0;
       /** Do a reverse dns lookup. */
-      virtual int32_t Resolve(socket *,in_addr a) = 0;
-      virtual int32_t Resolve(socket *,in6_addr& a) = 0;
+      virtual int Resolve(socket *,in_addr a) = 0;
+      virtual int Resolve(socket *,in6_addr& a) = 0;
       /** get listen port of asynchronous dns server. */
       virtual port_t GetResolverPort() = 0;
       /** Resolver thread ready for queries. */
@@ -184,17 +184,17 @@ namespace sockets
       /** Returns true if socket waiting for a resolve event. */
       virtual bool Resolving(socket *) = 0;
       /** Fetch unique trigger id. */
-      virtual int32_t TriggerID(socket *src) = 0;
+      virtual int TriggerID(socket *src) = 0;
       /** Subscribe socket to trigger id. */
-      virtual bool Subscribe(int32_t id, socket *dst) = 0;
+      virtual bool Subscribe(int id, socket *dst) = 0;
       /** Unsubscribe socket from trigger id. */
-      virtual bool Unsubscribe(int32_t id, socket *dst) = 0;
+      virtual bool Unsubscribe(int id, socket *dst) = 0;
       /** Execute OnTrigger for subscribed sockets.
       \param id Trigger ID
       \param data Data passed from source to destination
       \param erase Empty trigger id source and destination maps if 'true',
       Leave them in place if 'false' - if a trigger should be called many times */
-      virtual void Trigger(int32_t id, socket::TriggerData& data, bool erase = true) = 0;
+      virtual void Trigger(int id, socket::TriggerData& data, bool erase = true) = 0;
       /** Indicates that the handler runs under socket_thread. */
       virtual void SetSlave(bool x = true) = 0;
       /** Indicates that the handler runs under socket_thread. */
